+++ doc/whats-new.rst
@@ -21,6 +21,10 @@ v0.9.2 (unreleased)
 
 Enhancements
 ~~~~~~~~~~~~
+- When bottleneck version 1.1 or later is installed, use bottleneck for rolling
+  `var`, `argmin`, `argmax`, and `rank` computations. Also, `rolling.median`
+  now also accepts a `min_periods` argument (:issue:`1276`).
+  By `Joe Hamman <https://github.com/jhamman>`_.
 
 Bug fixes
 ~~~~~~~~~